# Cut verification/signing/proofs etc. time by half for 64bit machines