--- a/src/aggsig.rs
+++ b/src/aggsig.rs
@@ -392,6 +392,7 @@ impl Drop for AggSigContext {
 
 #[cfg(test)]
 mod tests {
+	extern crate chrono;
 	use super::{
 		add_signatures_single, export_secnonce_single, sign_single, verify_single, verify_batch,
 		AggSigContext, Secp256k1,
@@ -401,6 +402,7 @@ mod tests {
 	use rand::{thread_rng, Rng};
 	use crate::ContextFlag;
 	use crate::{AggSigPartialSignature, Message, Signature};
+	use crate::aggsig::tests::chrono::Utc;
 
 	#[test]
 	fn test_aggsig_multisig() {
@@ -508,6 +510,7 @@ mod tests {
 
 	#[test]
 	fn test_aggsig_batch() {
+		let nano_to_millis = 1.0 / 1_000_000.0;
 		let secp = Secp256k1::with_caps(ContextFlag::Full);
 
 		let mut sigs: Vec<Signature> = vec![];
@@ -531,7 +534,11 @@ mod tests {
 		}
 
 		println!("Verifying aggsig batch of 100");
+		let start = Utc::now().timestamp_nanos();
 		let result = verify_batch(&secp, &sigs, &msgs, &pub_keys);
+		let fin = Utc::now().timestamp_nanos();
+		let dur_ms = (fin - start) as f64 * nano_to_millis;
+		println!("{} signature batch validated in {}ms", 100, dur_ms);
 		assert!(result == true);
 	}