Taking the nice performance optimization for 64bit target from https://github.com/mimblewimble/rust-secp256k1-zkp/pull/71.

A few tests result on my desktop:

1. Schnorr Signature Batch Verification:

Before this optimization:
```
$ cargo test --release -- test_aggsig_batch --nocapture

running 1 test
Verifying aggsig batch of 100
100 signature batch validated in 7.218ms
test aggsig::tests::test_aggsig_batch ... ok
```

After this optimization:
```
$ cargo test --release -- test_aggsig_batch --nocapture

running 1 test
Verifying aggsig batch of 100
100 signature batch validated in 4.019ms
test aggsig::tests::test_aggsig_batch ... ok
```

2. Bullet Proof Batch Verification:
```
$ cargo test --release -- bench_bullet_proof_single_vs_multi --nocapture

running 1 test
--------
Comparing 1 Proofs
1 proofs single validated in 4.95ms
1 proofs batch validated in 4.9879999999999995ms
--------
Comparing 2 Proofs
2 proofs single validated in 10.259ms
2 proofs batch validated in 5.763999999999999ms
--------
Comparing 5 Proofs
5 proofs single validated in 24.759999999999998ms
5 proofs batch validated in 7.856ms
--------
Comparing 10 Proofs
10 proofs single validated in 48.291ms
10 proofs batch validated in 11.266ms
--------
Comparing 100 Proofs
100 proofs single validated in 482.11699999999996ms
100 proofs batch validated in 70.78099999999999ms
--------
Comparing 200 Proofs
200 proofs single validated in 972.313ms
200 proofs batch validated in 137.71ms
test pedersen::tests::bench_bullet_proof_single_vs_multi ... ok
```

After this optimization:
```
$ cargo test --release -- bench_bullet_proof_single_vs_multi --nocapture

running 1 test
--------
Comparing 1 Proofs
1 proofs single validated in 2.993ms
1 proofs batch validated in 2.8089999999999997ms
--------
Comparing 2 Proofs
2 proofs single validated in 5.795999999999999ms
2 proofs batch validated in 3.311ms
--------
Comparing 5 Proofs
5 proofs single validated in 14.263ms
5 proofs batch validated in 4.516ms
--------
Comparing 10 Proofs
10 proofs single validated in 28.397ms
10 proofs batch validated in 6.451ms
--------
Comparing 100 Proofs
100 proofs single validated in 283.887ms
100 proofs batch validated in 39.845ms
--------
Comparing 200 Proofs
200 proofs single validated in 568.4889999999999ms
200 proofs batch validated in 75.88ms
test pedersen::tests::bench_bullet_proof_single_vs_multi ... ok
```